API

IntervalArithmetic.BareInterval — Type

BareInterval{T<:NumTypes}

Interval type for guaranteed computation with interval arithmetic according to the IEEE Standard 1788-2015. Unlike Interval, this bare interval does not have decorations, is not a subtype of Real and errors on operations mixing BareInterval and Number.

Fields:

lo :: T

hi :: T

Constructor compliant with the IEEE Standard 1788-2015: bareinterval.

See also: Interval.

IntervalArithmetic.Decoration — Type

Decoration

Enumeration constant for the types of interval decorations described in Section 11.2 of the IEEE Standard 1788-2015:

com -> 4 (common): non-empty, continuous and bounded interval.

dac -> 3 (defined and continuous): non-empty and continuous interval.

def -> 2 (defined): non-empty interval.

trv -> 1 (trivial): meaningless interval.

ill -> 0 (ill-formed): not an interval (NaI).

IntervalArithmetic.Flavor — Type

Flavor{F}

A flavor defining how an interval behaves in edge cases. For instance, infinity may or not be considered part of unbounded intervals.

Some flavors F include:

:set_based (default): elements of an interval are real numbers. In particular, infinity is never part of an interval. This flavor is described and required in Part 2 of the IEEE Standard 1788-2015. Edge cases: - any unbounded interval does not contain infinity. - $[0, 0] / [0, 0] = \emptyset$. - $x / [0, 0] = \emptyset$ for any interval $x$. - $x \times [0, 0] = [0, 0]$ for any interval $x$.

:cset: elements of an interval are either real numbers, or $\pm \infty$, applying standard rule for arithmetic with infinity. Edge cases: - any unbounded interval contains infinity. - $[0, 0] / [0, 0] = [-\infty, \infty]$. - $x / [0, 0] = [-\infty, \infty]$ for any interval $x$. - $x \times [0, 0] = [-\infty, \infty]$ for any unbounded interval $x$.

Note

Currently only the flavor :set_based is supported and implemented.

Examples

julia> IntervalArithmetic.default_flavor()
IntervalArithmetic.Flavor{:set_based}()

julia> isempty_interval(bareinterval(Inf, Inf))
┌ Warning: invalid interval, empty interval is returned
└ @ IntervalArithmetic ~/work/IntervalArithmetic.jl/IntervalArithmetic.jl/src/intervals/construction.jl:202
true

julia> isempty_interval(bareinterval(0)/bareinterval(0))
true

julia> isempty_interval(bareinterval(1)/bareinterval(0))
true

julia> isempty_interval(bareinterval(-Inf, Inf)/bareinterval(0))
true

julia> isthinzero(bareinterval(0)*bareinterval(-Inf, Inf))
true

IntervalArithmetic.Interval — Type

Interval{T<:NumTypes} <: Real

Interval type for guaranteed computation with interval arithmetic according to the IEEE Standard 1788-2015. This structure combines a BareInterval together with a Decoration.

Fields:

bareinterval :: BareInterval{T}

decoration :: Decoration

isguaranteed :: Bool

Constructors compliant with the IEEE Standard 1788-2015:

interval

..

±

@I_str

See also: ±, .. and @I_str.

IntervalArithmetic.IntervalRounding — Type

IntervalRounding

Interval rounding type.

Available rounding types:

:fast (unsupported): rounding via prevfloat and nextfloat.

:tight: rounding via RoundingEmulator.jl.

:slow: rounding via setrounding.

:none: no rounding (non-rigorous numerics).

IntervalArithmetic.PowerMode — Type

PowerMode

Power mode type for ^.

Available mode types:

:fast (default): x ^ y is semantically equivalent to fastpow(x, y),

unless isthininteger(y) is true in which case it is semantically equivalent to fastpown(x, sup(y)).

:slow: x ^ y is semantically equivalent to pow(x, y), unless

isthininteger(y) is true in which case it is semantically equivalent to pown(x, sup(y)).

Base.:* — Method

*(x::BareInterval, y::BareInterval)
*(x::Interval, y::Interval)

Implement the mul function of the IEEE Standard 1788-2015 (Table 9.1).

Note

The behavior of * is flavor dependent for some edge cases.

Base.:+ — Method

+(x::BareInterval, y::BareInterval)
+(x::Interval, y::Interval)

Implement the add function of the IEEE Standard 1788-2015 (Table 9.1).

Base.:- — Method

-(x::BareInterval)
-(x::Interval)

Implement the neg function of the IEEE Standard 1788-2015 (Table 9.1).

Base.:- — Method

-(x::BareInterval, y::BareInterval)
-(x::Interval, y::Interval)

Implement the sub function of the IEEE Standard 1788-2015 (Table 9.1).

Base.:/ — Method

/(x::BareInterval, y::BareInterval)
/(x::Interval, y::Interval)

Implement the div function of the IEEE Standard 1788-2015 (Table 9.1).

Note

The behavior of / is flavor dependent for some edge cases.

Base.:== — Method

==(::BareInterval, ::Number)
==(::Number, ::BareInterval)
==(::Interval, ::Number)
==(::Number, ::Interval)

Test whether an interval is the singleton of a given number. In other words, the result is true if and only if the interval contains only that number.

Note

Comparison between intervals is purposely disallowed. Indeed, equality between non-singleton intervals has distinct properties, notably $x = y$ does not imply $x - y = 0$. See instead isequal_interval.

Base.:^ — Method

^(x::BareInterval, y::BareInterval)
^(x::Interval, y::Interval)

Compute the power of the positive real part of x by y. This function is not in the IEEE Standard 1788-2015. Its behaviour depend on the current PowerMode.

See also: pow and pown.

Examples

julia> setdisplay(:full);

julia> bareinterval(2, 3) ^ bareinterval(2)
BareInterval{Float64}(4.0, 9.0)

julia> interval(-1, 1) ^ interval(3)
Interval{Float64}(-1.0, 1.0, com)

julia> interval(-1, 1) ^ interval(-3)
Interval{Float64}(-Inf, Inf, trv)

Base.Math.acot — Method

acot(::BareInterval)
acot(::Interval)

Implement the acot function of the IEEE Standard 1788-2015 (Table 9.1).

Base.Math.acoth — Method

acoth(::BareInterval)
acoth(::Interval)

Implement the acoth function of the IEEE Standard 1788-2015 (Table 9.1).

Base.Math.cot — Method

cot(::BareInterval)
cot(::Interval)

Implement the cot function of the IEEE Standard 1788-2015 (Table 9.1).

Base.Math.coth — Method

coth(::BareInterval)
coth(::Interval)

Implement the coth function of the IEEE Standard 1788-2015 (Table 9.1).

Base.Math.csc — Method

csc(::BareInterval)
csc(::Interval)

Implement the csc function of the IEEE Standard 1788-2015 (Table 9.1).

Base.Math.csch — Method

csch(::BareInterval)
csch(::Interval)

Implement the csch function of the IEEE Standard 1788-2015 (Table 9.1).

Base.Math.hypot — Method

hypot(x, y)

Compute the hypotenuse.

Implement the hypot function of the IEEE Standard 1788-2015 (Table 10.5).

Base.Math.sec — Method

sec(::BareInterval)
sec(::Interval)

Implement the sec function of the IEEE Standard 1788-2015 (Table 9.1).

Base.Math.sech — Method

sech(::BareInterval)
sech(::Interval)

Implement the sech function of the IEEE Standard 1788-2015 (Table 9.1).

Base.abs — Method

abs(x::BareInterval)
abs(x::Interval)

Implement the abs function of the IEEE Standard 1788-2015 (Table 9.1).

Base.abs2 — Method

abs2(x::BareInterval)
abs2(x::Interval)

Implement the square absolute value.

Note

This function calls ^ internally, hence it depends on IntervalArithmetic.power_mode().

Base.acos — Method

acos(::BareInterval)
acos(::Interval)

Implement the acos function of the IEEE Standard 1788-2015 (Table 9.1).

Base.acosh — Method

acosh(::BareInterval)
acosh(::Interval)

Implement the acosh function of the IEEE Standard 1788-2015 (Table 9.1).

Base.asin — Method

asin(::BareInterval)
asin(::Interval)

Implement the asin function of the IEEE Standard 1788-2015 (Table 9.1).

Base.asinh — Method

asinh(::BareInterval)
asinh(::Interval)

Implement the asinh function of the IEEE Standard 1788-2015 (Table 9.1).

Base.atan — Method

atan(::BareInterval)
atan(::Interval)

Implement the atan function of the IEEE Standard 1788-2015 (Table 9.1).

Base.atan — Method

atan(::BareInterval, ::BareInterval)
atan(::Interval, ::Interval)

Implement the atan2 function of the IEEE Standard 1788-2015 (Table 9.1).

Base.atanh — Method

atanh(::BareInterval)
atanh(::Interval)

Implement the atanh function of the IEEE Standard 1788-2015 (Table 9.1).

Base.ceil — Method

ceil(x::BareInterval)
ceil(x::Interval)

Implement the ceil function of the IEEE Standard 1788-2015 (Table 9.1).

Base.cos — Method

cos(::BareInterval)
cos(::Interval)

Implement the cos function of the IEEE Standard 1788-2015 (Table 9.1).

Base.cosh — Method

cosh(::BareInterval)
cosh(::Interval)

Implement the cosh function of the IEEE Standard 1788-2015 (Table 9.1).

Base.floor — Method

floor(x::BareInterval)
floor(x::Interval)

Implement the floor function of the IEEE Standard 1788-2015 (Table 9.1).

Base.fma — Method

fma(x::BareInterval, y::BareInterval, z::BareInterval)
fma(x::Interval, y::Interval, z::Interval)

Implement the fma function of the IEEE Standard 1788-2015 (Table 9.1).

Base.hash — Method

hash(x::BareInterval, h::UInt)
hash(x::Interval, h::UInt)

Compute the integer hash code for an interval. Note that equality of intervals is given by isequal_interval rather than the == operator.

Base.inv — Method

inv(x::BareInterval)
inv(x::Interval)

Implement the recip function of the IEEE Standard 1788-2015 (Table 9.1).

Note

The behavior of inv is flavor dependent for some edge cases.

Base.max — Method

max(x::BareInterval, y::BareInterval)
max(x::Interval, y::Interval)

Implement the max function of the IEEE Standard 1788-2015 (Table 9.1).

Base.min — Method

min(x::BareInterval, y::BareInterval)
min(x::Interval, y::Interval)

Implement the min function of the IEEE Standard 1788-2015 (Table 9.1).

Base.muladd — Method

muladd(x::BareInterval, y::BareInterval z::BareInterval)
muladd(x::Interval, y::Interval z::Interval)

Implement the combined multiply-add; this is semantically equivalent to x * y + z.

Base.parse — Method

parse(T<:BareInterval, str)
parse(T<:Interval, str)

Create an interval according to the IEEE Standard 1788-2015. In contrast with constructors that do not use strings, this constructor guarantees that the returned interval tightly encloses the values described by the string, including numbers that have no exact float representation (e.g. 0.1).

Parse a string of the form "[a, b]_dec" as an Interval with decoration dec. If the decoration is not specified, it is computed based on the parsed interval. If the input is an invalid string, a warning is printed and NaI is returned. The parser is case unsensitive.

Examples of allowed string formats:

I"[1.33]" or I"1.33": the interval containing $1.33$.

I"[1.44, 2.78]": the interval $[1.44, 2.78]$.

I"[empty]": the empty interval.

I"[entire]" or I"[,]": the interval $[-\infty, \infty]$.

I"[3,]": the interval $[3, \infty]$.

I"6.42?2": the interval $[6.4, 6.44]$. The number after ? represents the uncertainty in the last digit; by default this value is 0.5. The direction of the uncertainty can be given by adding u or d at the end for the error to only go up or down respectively (e.g. I"4.5?5u" represents $[4.5, 5]$).

I"6.42?2e2": the interval $[642, 644]$.

I"3??u": the interval $[3, \infty]$.

I"3??u": the interval $[3, \infty]$.

I"3??": the interval $[-\infty, \infty]$.

For more details, see sections 9.7 and 12.11 of the IEEE Standard 1788-2015.

Examples

julia> setdisplay(:full);

julia> parse(BareInterval{Float64}, "[1, 2]")
BareInterval{Float64}(1.0, 2.0)

julia> parse(BareInterval{Float64}, "[1,]")
BareInterval{Float64}(1.0, Inf)

julia> parse(BareInterval{Float64}, "[,]")
BareInterval{Float64}(-Inf, Inf)

julia> parse(BareInterval{Float64}, "6.42?2e2")
BareInterval{Float64}(640.0, 644.0)

julia> parse(Interval{Float64}, "[1, 2]")
Interval{Float64}(1.0, 2.0, com)

julia> parse(Interval{Float64}, "[1, 2]_def")
Interval{Float64}(1.0, 2.0, def)

julia> parse(Interval{Float64}, "[1, 1e400]_com")
Interval{Float64}(1.0, Inf, dac)

Base.round — Method

round(x::BareInterval, [RoundingMode])
round(x::Interval, [RoundingMode])

Return an interval with the bounds of x rounded to an integer.

Implement the functions roundTiesToEven and roundTiesToAway of the IEEE Standard 1788-2015.

Base.sign — Method

sign(x::BareInterval)
sign(x::Interval)

Implement the sign function of the IEEE Standard 1788-2015 (Table 9.1).

Base.sin — Method

sin(::BareInterval)
sin(::Interval)

Implement the sin function of the IEEE Standard 1788-2015 (Table 9.1).

Base.sinh — Method

sinh(::BareInterval)
sinh(::Interval)

Implement the sinh function of the IEEE Standard 1788-2015 (Table 9.1).

Base.sqrt — Method

sqrt(x::BareInterval)
sqrt(x::Interval)

Implement the sqrt function of the IEEE Standard 1788-2015 (Table 9.1).

Base.tan — Method

tan(::BareInterval)
tan(::Interval)

Implement the tan function of the IEEE Standard 1788-2015 (Table 9.1).

Base.tanh — Method

tanh(::BareInterval)
tanh(::Interval)

Implement the tanh function of the IEEE Standard 1788-2015 (Table 9.1).

Base.trunc — Method

trunc(x::BareInterval)
trunc(x::Interval)

Implement the trunc function of the IEEE Standard 1788-2015 (Table 9.1).

IntervalArithmetic._bigequiv — Method

_bigequiv(x)

Create a BigFloat with the same underlying precision.

IntervalArithmetic._interval_infsup — Method

_interval_infsup(T<:NumTypes, a, b, [d::Decoration])

Internal constructor for intervals described by their lower and upper bounds, i.e. of the form $[a, b]$.

IntervalArithmetic._interval_midpoint — Method

_interval_midpoint(T<:NumTypes, m, r, d = com)

Internal constructor for intervals described by their midpoint and radius, i.e. of the form $m \pm r$.

IntervalArithmetic._round_expr — Method

_round_expr(ex::Expr, rounding_mode::RoundingMode)

Transforms a single expression by applying a rounding mode, e.g.

a + b into +(a, b, RoundDown)

sin(a) into sin(a, RoundDown)

IntervalArithmetic._unsafe_bareinterval — Function

_unsafe_bareinterval(T<:NumTypes, lo, hi)

Internal constructor which assumes that is_valid_interval(lo, hi) == true.

Danger

This constructor is not compliant with the IEEE Standard 1788-2015. Since misuse of this function can deeply corrupt code, its usage is strongly discouraged in favour of bareinterval.

IntervalArithmetic._unsafe_interval — Function

_unsafe_interval(bareinterval::BareInterval, decoration::Decoration, isguaranteed::Bool)

Internal constructor which assumes that bareinterval and its decoration decoration are compliant with the IEEE Standard 1788-2015.

Danger

This constructor is not compliant with the IEEE Standard 1788-2015. Since misuse of this function can deeply corrupt code, its usage is strongly discouraged in favour of interval.

IntervalArithmetic.atomic — Method

atomic(T<:Union{Rational,AbstractFloat}, x)

Create an interval according to the IEEE Standard 1788-2015. The returned Interval{T} always contains the value x; this is semantically equivalent to parse(Interval{T}, string(x)) if x is a Number.

Examples

julia> setdisplay(:full);

julia> x = IntervalArithmetic.atomic(Float64, 0.1)
Interval{Float64}(0.09999999999999999, 0.1, com)

julia> in_interval(1//10, IntervalArithmetic.atomic(Float64, 0.1))
true

julia> IntervalArithmetic.atomic(Float64, 0.3)
Interval{Float64}(0.3, 0.30000000000000004, com)

julia> in_interval(3//10, IntervalArithmetic.atomic(Float64, 0.3))
true

IntervalArithmetic.bareinterval — Method

bareinterval([T<:Union{Rational,AbstractFloat}=default_numtype()], a, b)

Create the bare interval $[a, b]$ according to the IEEE Standard 1788-2015. The validity of the interval is checked by is_valid_interval: if true then a BareInterval{T} is constructed, otherwise an empty interval is returned.

Danger

Nothing is done to compensate for the fact that floating point literals are rounded to the nearest when parsed (e.g. 0.1). In such cases, parse the string containing the desired value to ensure its tight enclosure.

See also: interval, ±, .. and @I_str.

Examples

julia> setdisplay(:full);

julia> bareinterval(1//1, π)
BareInterval{Rational{Int64}}(1//1, 85563208//27235615)

julia> bareinterval(Rational{Int32}, 1//1, π)
BareInterval{Rational{Int32}}(1//1, 85563208//27235615)

julia> bareinterval(1, π)
BareInterval{Float64}(1.0, 3.1415926535897936)

julia> bareinterval(BigFloat, 1, π)
BareInterval{BigFloat}(1.0, 3.141592653589793238462643383279502884197169399375105820974944592307816406286233)

IntervalArithmetic.bisect — Method

bisect(x, α=0.5)
bisect(x, i, α=0.5)

Split an interval x at a relative position α, where α = 0.5 corresponds to the midpoint.

Split the i-th component of a vector x at a relative position α, where α = 0.5 corresponds to the midpoint.

IntervalArithmetic.bounds — Method

bounds(x)

Bounds of x given as a tuple. Unlike inf, this function does not normalize the infimum of the interval.

See also: inf, sup, mid, diam, radius and midradius.

IntervalArithmetic.cancelminus — Method

cancelminus(x, y)

Compute the unique interval z such that y + z == x.

The result is decorated by at most trv (Section 11.7.1).

Implement the cancelMinus function of the IEEE Standard 1788-2015 (Section 9.2).

IntervalArithmetic.cancelplus — Method

cancelplus(x, y)

Compute the unique interval z such that y - z == x; this is semantically equivalent to cancelminus(x, -y).

The result is decorated by at most trv (Section 11.7.1).

Implement the cancelPlus function of the IEEE Standard 1788-2015 (Section 9.2).

IntervalArithmetic.contains_infinity — Method

contains_infinity([F::Flavor=default_flavor()], x::BareInterval)

For the given flavor F, test whether x contains infinity.

IntervalArithmetic.default_flavor — Method

default_flavor()

Return the default flavor used to handle edge cases.

IntervalArithmetic.default_numtype — Method

default_numtype()

Return the default bound type used in promote_numtype. By default, default_numtype() is set to Float64. It can be modified by redefining the function, however it should be set to a concrete subtype of NumTypes.

Examples

julia> IntervalArithmetic.default_numtype() = Float32

julia> typeof(interval(1, 2))
Interval{Float32}

julia> typeof(interval(1, big(2)))
Interval{BigFloat}

julia> IntervalArithmetic.default_numtype() = Float64

julia> typeof(interval(1, 2))
Interval{Float64}

julia> typeof(interval(1, big(2)))
Interval{BigFloat}

IntervalArithmetic.diam — Method

diam(x)

Diameter of x. If x is complex, then the diameter is the maximum diameter between its real and imaginary parts.

Implement the wid function of the IEEE Standard 1788-2015 (Table 9.2).

See also: inf, sup, bounds, mid, radius and midradius.

IntervalArithmetic.dist — Method

dist(x, y)

Distance between x and y.

IntervalArithmetic.div_by_thin_zero — Method

div_by_thin_zero([F::Flavor=default_flavor()], x::BareInterval)

For the given flavor F, divide x by the interval containing only $0$.

IntervalArithmetic.emptyinterval — Method

emptyinterval(T=default_numtype())

Create an empty interval. This interval is an exception to the fact that the lower bound is larger than the upper one.

Implement the empty function of the IEEE Standard 1788-2015 (Section 10.5.2).

IntervalArithmetic.entireinterval — Method

entireinterval(T=default_numtype())

Create an interval representing the entire real line, or the entire complex plane if T is complex.

Note

Depending on the flavor, infinity may or may not be considered part of the interval.

Implement the entire function of the IEEE Standard 1788-2015 (Section 10.5.2).

IntervalArithmetic.extended_div — Method

extended_div(x, y)

Two-output division.

Implement the mulRevToPair function of the IEEE Standard 1788-2015 (Section 10.5.5).

IntervalArithmetic.fastpow — Method

fastpow(x, y)

A faster implementation of pow(x, y), at the cost of maybe returning a slightly larger interval.

IntervalArithmetic.fastpown — Method

fastpown(x, n)

A faster implementation of pown(x, y), at the cost of maybe returning a slightly larger interval.

IntervalArithmetic.hull — Method

hull(x, y)

Return the interval hull of the intervals x and y, considered as (extended) sets of real numbers, i.e. the smallest interval that contains all of x and y.

The result is decorated by at most trv (Section 11.7.1).

Implement the convexHull function of the IEEE Standard 1788-2015 (Section 9.3).

IntervalArithmetic.in_interval — Method

in_interval(x, y)

Test whether x is an element of y.

Implement the isMember function of the IEEE Standard 1788-2015 (Section 10.6.3).

IntervalArithmetic.inf — Method

inf(x)

Lower bound, or infimum, of x. For a zero AbstractFloat lower bound, a negative zero is returned.

Implement the inf function of the IEEE Standard 1788-2015 (Table 9.2).

See also: sup, bounds, mid, diam, radius and midradius.

IntervalArithmetic.interiordiff! — Method

interiordiff(x, y)

In-place version of interiordiff.

IntervalArithmetic.interiordiff — Method

interiordiff(x, y)

Remove the interior of y from x. If x and y are vectors, then they are treated as multi-dimensional intervals.

IntervalArithmetic.intersect_interval — Method

intersect_interval(x, y)

Returns the intersection of the intervals x and y, considered as (extended) sets of real numbers. That is, the set that contains the points common in x and y.

The result is decorated by at most trv (Section 11.7.1).

Implement the intersection function of the IEEE Standard 1788-2015 (Section 9.3).

IntervalArithmetic.interval — Method

interval([T<:Union{Rational,AbstractFloat}=default_numtype()], a, b, d = com)

Create the interval $[a, b]$ according to the IEEE Standard 1788-2015. The validity of the interval is checked by is_valid_interval: if true then an Interval{T} is constructed, otherwise an NaI (Not an Interval) is returned.

Danger

Nothing is done to compensate for the fact that floating point literals are rounded to the nearest when parsed (e.g. 0.1). In such cases, parse the string containing the desired value to ensure its tight enclosure.

See also: ±, .. and @I_str.

Examples

julia> setdisplay(:full);

julia> interval(1//1, π)
Interval{Rational{Int64}}(1//1, 85563208//27235615, com)

julia> interval(Rational{Int32}, 1//1, π)
Interval{Rational{Int32}}(1//1, 85563208//27235615, com)

julia> interval(1, π)
Interval{Float64}(1.0, 3.1415926535897936, com)

julia> interval(BigFloat, 1, π)
Interval{BigFloat}(1.0, 3.141592653589793238462643383279502884197169399375105820974944592307816406286233, com)

IntervalArithmetic.is_valid_interval — Method

is_valid_interval([F::Flavor=default_flavor()], a::Real, b::Real)

For the given flavor F, test whether $[a, b]$ is a valid interval.

IntervalArithmetic.isatomic — Method

isatomic(x)

Test whether x is unable to be split. This occurs if the interval is empty, or if its lower and upper bounds are equal, or if the bounds are consecutive floating-point numbers.

IntervalArithmetic.isbounded — Method

isbounded(x)

Test whether x is empty or has finite bounds.

IntervalArithmetic.iscommon — Method

iscommon(x)

Test whether x is not empty and bounded.

Note

This is does not take into consideration the decoration of the interval.

IntervalArithmetic.isdisjoint_interval — Method

isdisjoint_interval(x, y)

Test whether x and y have no common elements.

Implement the disjoint function of the IEEE Standard 1788-2015 (Table 9.3).

IntervalArithmetic.isempty_interval — Method

isempty_interval(x)

Test whether x contains no elements.

Implement the isEmpty function of the IEEE Standard 1788-2015 (Section 10.6.3).

IntervalArithmetic.isentire_interval — Method

isentire_interval(x)

Test whether x is the entire real line.

Implement the isEntire function of the IEEE Standard 1788-2015 (Section 10.6.3).

IntervalArithmetic.isequal_interval — Method

isequal_interval(x, y)

Test whether x and y are identical.

Implement the equal function of the IEEE Standard 1788-2015 (Table 9.3).

IntervalArithmetic.isguaranteed — Method

isguaranteed(x::BareInterval)
isguaranteed(x::Interval)
isguaranteed(x::Complex{<:Interval})

Test whether the interval is not guaranteed to encompass all possible numerical errors. This happens whenever an Interval is constructed using convert(::Type{<:Interval}, ::Real), which may occur implicitly when mixing intervals and Real types.

Since conversion between BareInterval and Number is prohibited, this implies that isguaranteed(::BareInterval) == true.

In the case of a complex interval x, this is semantically equivalent to isguaranteed(real(x)) & isguaranteed(imag(x)).

Examples

julia> isguaranteed(bareinterval(1))
true

julia> isguaranteed(interval(1))
true

julia> isguaranteed(convert(Interval{Float64}, 1))
false

julia> isguaranteed(interval(1) + 0)
false

IntervalArithmetic.isinterior — Method

isinterior(x, y)

Test whether x is in the interior of y.

Implement the interior function of the IEEE Standard 1788-2015 (Table 9.3).

See also: isstrictsubset.

IntervalArithmetic.isnai — Method

isnai(x)

Test whether x is an NaI (Not an Interval).

IntervalArithmetic.isstrictless — Method

isstrictless(x, y)

Test whether inf(x) < inf(y) and sup(x) < sup(y), where < is replaced by ≤ for infinite values.

Implement the strictLess function of the IEEE Standard 1788-2015 (Table 10.3).

IntervalArithmetic.isstrictsubset — Method

isstrictsubset(x, y)

Test whether x is a strict subset of y. If x and y are intervals, this is semantically equivalent to isinterior(x, y). If x and y are vectors, at least one component of x must be in the interior of y.

See also: isinterior.

IntervalArithmetic.issubset_interval — Method

issubset_interval(x, y)

Test whether x is contained in y.

Implement the subset function of the IEEE Standard 1788-2015 (Table 9.3).

IntervalArithmetic.isthin — Method

isthin(x, y)

Test whether x contains only y.

IntervalArithmetic.isthin — Method

isthin(x)

Test whether x contains only a real.

Implement the isSingleton function of the IEEE Standard 1788-2015 (Table 9.3).

IntervalArithmetic.isthininteger — Method

isthininteger(x)

Test whether x contains only an integer.

IntervalArithmetic.isthinone — Method

isthinone(x)

Test whether x contains only one.

IntervalArithmetic.isthinzero — Method

isthinzero(x)

Test whether x contains only zero.

IntervalArithmetic.isunbounded — Method

isunbounded(x)

Test whether x is not empty and has infinite bounds.

IntervalArithmetic.isweakless — Method

isweakless(x, y)

Test whether inf(x) ≤ inf(y) and sup(x) ≤ sup(y), where < is replaced by ≤ for infinite values.

Implement the less function of the IEEE Standard 1788-2015 (Table 10.3).

IntervalArithmetic.mag — Method

mag(x)

Magnitude of x.

Implement the mag function of the IEEE Standard 1788-2015 (Table 9.2).

See also: mig.

IntervalArithmetic.mid — Method

mid(x)

Midpoint of x.

Implement the mid function of the IEEE Standard 1788-2015 (Table 9.2).

See also: inf, sup, bounds, diam, radius and midradius.

IntervalArithmetic.midradius — Method

midradius(x)

Midpoint and radius of x.

Function required by the IEEE Standard 1788-2015 in Section 10.5.9 for the set-based flavor.

See also: inf, sup, bounds, mid, mid and radius.

IntervalArithmetic.mig — Method

mig(x)

Mignitude of x.

Implement the mig function of the IEEE Standard 1788-2015 (Table 9.2).

See also: mag.

IntervalArithmetic.mince! — Method

mince!(v, x, n)

In-place version of mince.

IntervalArithmetic.mince — Method

mince(x, n)

Split an interval x in n intervals of the same diameter.

Split the i-th component of a vector x in n[i] intervals of the same diameter; n can be a tuple of integers, or a single integer in which case the same n is used for all the components of x.

IntervalArithmetic.nai — Method

nai(T=default_numtype())

Create an NaI (Not an Interval).

IntervalArithmetic.numtype — Method

numtype(T)

Return the bounds type of the interval.

Examples

julia> IntervalArithmetic.numtype(interval(1, 2))
Float64

julia> IntervalArithmetic.numtype(interval(Float32, 1, 2))
Float32

IntervalArithmetic.overlap — Method

overlap(x::BareInterval, y::BareInterval)
overlap(x::Interval, y::Interval)

Implement the overlap function of the IEEE Standard 1788-2015 (Table 10.7).

IntervalArithmetic.pow — Method

pow(x, y)

Compute the power of the positive real part of x by y. In particular, even if y is a thin integer, this is not equivalent to pown(x, sup(y)).

Implement the pow function of the IEEE Standard 1788-2015 (Table 9.1).

See also: pown.

Examples

julia> setdisplay(:full);

julia> pow(bareinterval(2, 3), bareinterval(2))
BareInterval{Float64}(4.0, 9.0)

julia> pow(interval(-1, 1), interval(3))
Interval{Float64}(0.0, 1.0, trv)

julia> pow(interval(-1, 1), interval(-3))
Interval{Float64}(1.0, Inf, trv)

IntervalArithmetic.pown — Method

pown(x, n)

Implement the pown function of the IEEE Standard 1788-2015 (Table 9.1).

Examples

julia> setdisplay(:full);

julia> pown(bareinterval(2, 3), 2)
BareInterval{Float64}(4.0, 9.0)

julia> pown(interval(-1, 1), 3)
Interval{Float64}(-1.0, 1.0, com)

julia> pown(interval(-1, 1), -3)
Interval{Float64}(-Inf, Inf, trv)

IntervalArithmetic.precedes — Method

precedes(x, y)

Test whether any element of x is lesser or equal to every elements of y.

Implement the precedes function of the IEEE Standard 1788-2015 (Table 10.3).

IntervalArithmetic.promote_numtype — Method

promote_numtype(T, S)

Return the bound type used to construct intervals. The bound type is given by promote_type(T, S) if T or S is a Rational or an AbstractFloat; except when T is a Rational{R} and S is an AbstractIrrational (or vice-versa), in which case the bound type is given by Rational{promote_type(R, Int64)}. In all other cases, the bound type is given by promote_type(default_numtype(), T, S).

IntervalArithmetic.radius — Method

radius(x)

Radius of x, such that issubset_interval(x, mid(x) ± radius(x)). If x is complex, then the radius is the maximum radius between its real and imaginary parts.

Implement the rad function of the IEEE Standard 1788-2015 (Table 9.2).

See also: inf, sup, bounds, mid, diam and midradius.

IntervalArithmetic.rootn — Method

rootn(x::BareInterval, n::Integer)

Compute the real n-th root of x.

Implement the rootn function of the IEEE Standard 1788-2015 (Table 9.1).

IntervalArithmetic.setdisplay — Function

setdisplay(format::Symbol; decorations::Bool, sigdigits::Int)

Change the format used by show to display intervals.

Possible options:

format can be:

:infsup: display intervals as [a, b].

:midpoint: display intervals as m ± r.

:full: display interval bounds entirely, ignoring sigdigits.

decorations: display the decorations or not.

sigdigits: number (greater or equal to 1) of significant digits to display.

Initially, the display options are set to setdisplay(:infsup; decorations = false, sigdigits = 6). If any of the three argument format, decorations and sigdigits is omitted, then their value is left unchanged.

Examples

julia> setdisplay(:infsup; decorations = true, sigdigits = 6) # default display options
Display options:
  - format: infsup
  - decorations: true
  - significant digits: 6

julia> x = interval(0.1, 0.3)
[0.0999999, 0.300001]_com

julia> setdisplay(:full)
Display options:
  - format: full
  - decorations: true
  - significant digits: 6 (ignored)

julia> x
Interval{Float64}(0.1, 0.3, com)

julia> setdisplay(:infsup; sigdigits = 3)
Display options:
  - format: infsup
  - decorations: true
  - significant digits: 3

julia> x
[0.0999, 0.301]_com

julia> setdisplay(; decorations = false)
Display options:
  - format: infsup
  - decorations: false
  - significant digits: 3

julia> x
[0.0999, 0.301]

IntervalArithmetic.strictprecedes — Method

strictprecedes(x, y)

Test whether any element of x is strictly lesser than every elements of y.

Implement the strictPrecedes function of the IEEE Standard 1788-2015 (Table 10.3).

IntervalArithmetic.sup — Method

sup(x)

Upper bound, or supremum, of x.

Implement the sup function of the IEEE Standard 1788-2015 (Table 9.2).

See also: inf, bounds, mid, diam, radius and midradius.

IntervalArithmetic.zero_times_infinity — Method

zero_times_infinity([F::Flavor=default_flavor()], T<:NumTypes)

For the given flavor F, return $0 \times \infty$ as an instance of type T.

IntervalArithmetic.@I_str — Macro

I"str"

Create an interval by parsing the string "str"; this is semantically equivalent to parse(Interval{default_numtype()}, "str").

Examples

julia> setdisplay(:full);

julia> I"[3, 4]"
Interval{Float64}(3.0, 4.0, com)

julia> I"0.1"
Interval{Float64}(0.09999999999999999, 0.1, com)

julia> in_interval(1//10, I"0.1")
true

IntervalArithmetic.@interval — Macro

@interval(T, expr)
@interval(T, expr1, expr2)

Walk through an expression and wrap each argument of functions with the internal constructor atomic.

Examples

julia> setdisplay(:full);

julia> @macroexpand @interval Float64 sin(1)
:(sin(IntervalArithmetic.atomic(Float64, 1)))

julia> @interval Float64 sin(1) exp(1)
Interval{Float64}(0.8414709848078965, 2.7182818284590455, com)

IntervalArithmetic.@round — Macro

@round(T, ex1, ex2)

Macro for internal use that creates an interval by rounding down ex1 and rounding up ex2. Each expression may consist of only a single operation that needs rounding, e.g. a.lo + b.lo or sin(a.lo). It also handles min(...) and max(...), where the arguments are each themselves single operations.

The macro uses the internal _round_expr function to transform e.g. a + b into +(a, b, RoundDown).

IntervalArithmetic.NumTypes — Type

NumTypes

Constant for the supported types of interval bounds. This is set to Union{Rational,AbstractFloat}.